Update parameter type of gamma to float.
Fixed wrong parameter type of gamma even though it isn't impact
the test.

JIRA: IAHWC-55
Test: Build and run colorcorrection_autotest, it can get "PASSED"
      result as before.

+++ tests/autotests/colorcorrection_autotest.cpp
@@ -118,7 +118,7 @@ class HotPlugEventCallback : public hwcomposer::DisplayHotPlugEventCallback {
     }
   }
 
-  void SetGamma(uint32_t red, uint32_t green, uint32_t blue) {
+  void SetGamma(float red, float green, float blue) {
     hwcomposer::ScopedSpinLock lock(spin_lock_);
     PopulateConnectedDisplays();
